created _constants folder and renamed type file to user.types.js
