+++ electron/launcher/src/_constants/user.types.js
